con-b-o-change fix marketState after btm-reconnect

+++ core/src/main/java/com/bitplay/arbitrage/ArbitrageService.java
@@ -2052,7 +2052,7 @@ public class ArbitrageService {
     }
 
     public Long getLastTradeId() {
-        return tradeId != null ? tradeId.longValue() : fplayTradeRepository.getLastId();
+        return tradeId != null ? tradeId : fplayTradeRepository.getLastId();
     }
 
     public Long getLastInProgressTradeId() {
